Split TrackedInputContext implementations in their separate modules
